fix: separate the redis logic to its own module
